Adding example bullet code

+++ Phasor.JS example/public/src/game.js
@@ -1,12 +1,61 @@
+// import { Bullet } from './bullet.js'
 // Linted with standardJS - https://standardjs.com/
 
 // Initialize the Phaser Game object and set default game window size
 const game = new Phaser.Game(800, 600, Phaser.AUTO, '', {
     preload: preload,
     create: create,
-    update: update
+    update: update,
+    render: render
 })
 
+// var Bullet = new Phaser.Class({
+
+//     Extends: Phaser.GameObjects.Image,
+
+//     initialize:
+
+//     // Bullet Constructor
+//         function Bullet(scene) {
+//         Phaser.GameObjects.Image.call(this, scene, 0, 0, 'bullet');
+//         this.speed = 1;
+//         this.born = 0;
+//         this.direction = 0;
+//         this.xSpeed = 0;
+//         this.ySpeed = 0;
+//         this.setSize(12, 12, true);
+//     },
+
+//     // Fires a bullet from the player to the reticle
+//     fire: function(shooter, target) {
+//         this.setPosition(shooter.x, shooter.y); // Initial position
+//         this.direction = Math.atan((target.x - this.x) / (target.y - this.y));
+
+//         // Calculate X and y velocity of bullet to moves it from shooter to target
+//         if (target.y >= this.y) {
+//             this.xSpeed = this.speed * Math.sin(this.direction);
+//             this.ySpeed = this.speed * Math.cos(this.direction);
+//         } else {
+//             this.xSpeed = -this.speed * Math.sin(this.direction);
+//             this.ySpeed = -this.speed * Math.cos(this.direction);
+//         }
+
+//         this.rotation = shooter.rotation; // angle bullet with shooters rotation
+//         this.born = 0; // Time since new bullet spawned
+//     },
+
+//     // Updates the position of the bullet each cycle
+//     update: function(time, delta) {
+//         this.x += this.xSpeed * delta;
+//         this.y += this.ySpeed * delta;
+//         this.born += delta;
+//         if (this.born > 1800) {
+//             this.setActive(false);
+//             this.setVisible(false);
+//         }
+//     }
+
+// });
 // Declare shared variables at the top so all methods can access them
 let score = 0
 let scoreText
@@ -14,10 +63,19 @@ let platforms
 let diamonds
 let cursors
 let player
+let hazard
+var plat_x
+var plat_y
+var enemy1
+var tween1
+var brick
 var text;
 var timedEvent;
 var sky
 var json_parsed
+var goomba
+var walking_goomba
+var state = 3
 
 function preload() {
     // Load & Define our game assets
@@ -29,14 +87,20 @@ function preload() {
     game.load.audio("mario_die", './assets/smb_mariodie.wav')
     game.load.spritesheet("spike", "./assets/spike.png", 32, 32)
     game.load.spritesheet("brick", "./assets/brick.png", 32, 32)
-
+        //!~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
     game.load.text("Json_test", "./data/test.json")
+        //!~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    game.load.audio('coin_sound', './assets/smb_coin.wav')
+    game.load.audio('brick_sound', './assets/smb_breakblock.wav')
+
+    game.load.spritesheet('blue_goomba', './assets/bluegoomba.png', 32, 32)
 }
 
 function create() {
+    //!
     json_parsed = JSON.parse(game.cache.getText("Json_test"))
-    console.log(json_parsed)
-
+    console.log("Json file: ", json_parsed)
+        //!
 
     //  We're going to be using physics, so enable the Arcade Physics system
     game.physics.startSystem(Phaser.Physics.ARCADE)
@@ -65,12 +129,20 @@ function create() {
     ground.body.immovable = true
 
     //  Now let's create two ledges
-    let ledge = platforms.create(400, 450, 'ground')
-    ledge.body.immovable = true
 
+    //! ~~~~~ Parsing a json file and creating platforms ~~~~~
+    var platform_locations = json_parsed.Platforms
+    console.log("Platforms: ", platform_locations)
+    var ledge
+    for (var i = 0; i < platform_locations.length; i++) {
+        plat_x = platform_locations[i].x
+        plat_y = platform_locations[i].y
 
-    ledge = platforms.create(-75, 350, 'ground')
-    ledge.body.immovable = true
+        console.log("Creating platform at: ", plat_x, plat_y)
+        ledge = platforms.create(plat_x, plat_y, 'ground')
+        ledge.body.immovable = true
+    }
+    //! ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
     // The player and its settings
     player = game.add.sprite(32, game.world.height - 150, 'woof')
@@ -79,8 +151,8 @@ function create() {
     game.physics.arcade.enable(player)
 
     //  Player physics properties. Give the little guy a slight bounce.
-    player.body.bounce.y = 0.2
-    player.body.gravity.y = 980
+    player.body.bounce.y = 0
+    player.body.gravity.y = 990
     player.body.collideWorldBounds = true
 
     //  Our two animations, walking left and right.
@@ -93,14 +165,20 @@ function create() {
     //  Enable physics for any object that is created in this group
     diamonds.enableBody = true
 
-    //  Create 12 diamonds evenly spaced apart
-    for (var i = 0; i < 12; i++) {
-        const diamond = diamonds.create(i * 70, 0, 'diamond')
+    //! ~~~~~ CHANGE TO PARSE FROM JSON FILE ~~~~~
+    var diamound_location = json_parsed.Diamounds
+    for (var i = 0; i < diamound_location.length; i++) {
+        var diamound_x = diamound_location[i].x
+        var diamound_y = diamound_location[i].y
+
+        const diamond = diamonds.create(diamound_x, diamound_y, 'diamond')
 
         //  Drop em from the sky and bounce a bit
         diamond.body.gravity.y = 1000
         diamond.body.bounce.y = 0.3 + Math.random() * 0.2
     }
+    //! ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
 
     //  Create the score text
     scoreText = game.add.text(16, 16, '', { fontSize: '32px', fill: '#000' })
@@ -135,9 +213,9 @@ function create() {
 
     tween1 = game.add.tween(enemy1)
     tween1.loop = -1
-    tween1.to({ x: 350, y: 300 }, 2000, null, true, 0, loop = 100, true)
+    tween1.to({ x: 350, y: 300 }, 2000, null, true, 0, 100000, true)
 
-    console.log(this);
+    console.log("Game object: ", this);
     //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
     //~~~~~ Demo of birck ~~~~~
@@ -150,6 +228,18 @@ function create() {
     game.world.setBounds(0, 0, 8000, 600)
     game.camera.follow(player);
     //~~~~~~~~~~~~~~~~~~~~~~~~~
+
+    //! ~~~~~ Moving goomba w/ animations ~~~~~
+    goomba = hazard.create(1000, 368, 'blue_goomba')
+    goomba.animations.add('walk', [2, 1, 2, 0], 4, true)
+    goomba.animations.play('walk')
+    goomba.enableBody = true
+    goomba.body.gravity.y = 1000
+
+    walking_goomba = game.add.tween(goomba)
+    walking_goomba.loop = -1
+    walking_goomba.to({ x: 1250, y: 368 }, 2000, null, true, 0, 1000000, true)
+        //! ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 }
 
 function update() {
@@ -178,10 +268,18 @@ function update() {
         player.animations.stop()
     }
 
-    //  This allows the player to jump!
+    //! ~~~~~ Double jump demo ~~~~~
+    const jump_pressed = game.input.keyboard.justPressed(cursors.up)
+    var jump = 0
+    var duration = cursors.up.duration
+        //console.log("Jump pressed: ", jump_pressed)
     if (cursors.up.isDown && player.body.touching.down) {
+
         player.body.velocity.y = -500
+
     }
+    //! ~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
     // Show an alert modal when score reaches 120
     if (score === 120) {
         alert('You win!')
@@ -189,9 +287,20 @@ function update() {
     }
 }
 
+function render() {
+    this.game.debug.text(`Debugging Phaser ${Phaser.VERSION}`, 200, 20, 'yellow', 'Segoe UI');
+    this.game.debug.cameraInfo(this.game.camera, 200, 32);
+    this.game.debug.spriteInfo(player, 500, 32);
+}
+
 function collectDiamond(player, diamond) {
-    // Removes the diamond from the screen
+
+    console.log("Collect diamound ", player, diamond)
+    var coin_sound = game.add.audio('coin_sound')
+    coin_sound.play()
+        // Removes the diamond from the screen
     diamond.kill()
+
 
     //  And update the score
     score += 10
@@ -199,6 +308,7 @@ function collectDiamond(player, diamond) {
 }
 
 var tick = function() {
+    console.log("Tick: ", this)
     this.timeLimit--;
     var minutes = Math.floor(this.timeLimit / 60);
     var seconds = this.timeLimit - (minutes * 60);
@@ -224,16 +334,23 @@ var outofTime = function() {
 }
 
 function kill_mario(player, enemy) {
-    player.kill();
+    if (state >= 2) {
+        player.position.x = player.position.x - 25
+        state--
+        console.log("state: ", state)
+        return
+    } else {
 
-    var die_noise = game.add.audio("mario_die");
-    die_noise.play();
+        player.kill();
 
-    alert("Game over");
+        var die_noise = game.add.audio("mario_die");
+        die_noise.play();
 
-    location.reload();
+        alert("Game over");
+
+        location.reload();
+    }
 }
-
 
 function brick_break(player, block) {
     //Only break the brick when the player is below 
@@ -254,8 +371,10 @@ function brick_break(player, block) {
     } else if (player_x < block_x - 16) {
         return
     } else {
-        block.kill()
-    }
 
+        block.kill()
+        var break_sound = game.add.audio('brick_sound')
+        break_sound.play()
+    }
 
 }
